Refactor AddPanel with TextAreaEditor
