+++ packages/ui/src/components/TextAreaEditor.svelte
