+++ packages/ui/src/index.ts
@@ -58,7 +58,7 @@ export { default as PopupMenu } from './components/PopupMenu.svelte'
 export { default as SelectPopup } from './components/SelectPopup.svelte'
 export { default as ColorPopup } from './components/ColorPopup.svelte'
 export { default as TextArea } from './components/TextArea.svelte'
-export { default as TextAreaEditor } from './components/TextAriaEditor.svelte'
+export { default as TextAreaEditor } from './components/TextAreaEditor.svelte'
 export { default as Section } from './components/Section.svelte'
 export { default as DatePickerPopup } from './components/calendar/DatePickerPopup.svelte'
 export { default as DatePicker } from './components/calendar/DatePicker.svelte'
